Move the constant values into Fallout.Core
Core is the innermost project, so it is where Fallout's own fixed values belong
(#584). Everything now reads them from there.

Split Constants.cs on the line between data and behaviour:

- Fallout.Core/Constants.cs — the ~31 const values: directory and file names,
  parameter names, environment variable keys, package ids, project URLs.
  Adds FalloutGlobalToolPackageId, which UpdateNotificationAttribute previously
  carried as a literal.
- Fallout.Build.Shared/FalloutPaths.cs — the ~19 helpers that derive a path or
  read the file system. These cannot follow: they need AbsolutePath,
  EnvironmentInfo and the file system, and Core references nothing.

Fallout.Build.Shared now references Core, so the values have one home and the
helpers consume them from it.

Core gains netstandard2.0 for this. Fallout.SourceGenerators (Roslyn) and
Fallout.MSBuildTasks (net472) both consume these values and cannot load a
netstandard2.1 assembly. Core references nothing and uses no 2.1-only API, so
the 2.0 build is the same source; netstandard2.1 stays so 2.1+ consumers keep
the richer asset and future Core code can use 2.1 APIs behind an #if. The reason
and the exit condition are in a comment on the csproj.

Call sites: helper uses moved to FalloutPaths, either qualified or via an added
`using static`. No behaviour change — every value and every helper body is
identical to before.

Verified: solution builds clean, Core emits all three assets, full test suite
passes. Three GitRepositoryWorktreeSpecs failures are pre-existing on this
checkout (its `origin` is the fork, so the parsed identifier is
ChrisonSimtian/Fallout rather than Fallout-build/Fallout) and reproduce
unchanged at upstream/main.
