+++ src/Fallout.Build/FalloutBuild.Statics.cs
@@ -8,6 +8,7 @@ using Fallout.Common.IO;
 using Fallout.Common.Utilities;
 using Fallout.Common.Utilities.Collections;
 using static Fallout.Common.Constants;
+using static Fallout.Common.FalloutPaths;
 
 namespace Fallout.Common;
