+++ src/Fallout.Cli/BuildScaffolder.cs
@@ -6,6 +6,7 @@ using Fallout.Common;
 using Fallout.Common.IO;
 using Fallout.Common.Utilities;
 using static Fallout.Common.Constants;
+using static Fallout.Common.FalloutPaths;
 using static Fallout.Common.EnvironmentInfo;
 using static Fallout.Common.Tooling.ProcessTasks;
 using static Fallout.Common.Utilities.TemplateUtility;
