+++ src/Fallout.Cli/CakeConverter.cs
@@ -9,6 +9,7 @@ using Fallout.Common.Tooling;
 using Fallout.Common.Utilities;
 using Fallout.Cli.Rewriting.Cake;
 using static Fallout.Common.Constants;
+using static Fallout.Common.FalloutPaths;
 using static Fallout.Common.EnvironmentInfo;
 
 namespace Fallout.Cli;
