+++ src/Fallout.Cli/Commands/SecretsCommand.cs
@@ -7,6 +7,7 @@ using Fallout.Common.IO;
 using Fallout.Common.Utilities;
 using Fallout.Common.Utilities.Collections;
 using static Fallout.Common.Constants;
+using static Fallout.Common.FalloutPaths;
 using static Fallout.Common.Utilities.EncryptionUtility;
 
 namespace Fallout.Cli.Commands;
