+++ src/Fallout.Core/Fallout.Core.csproj
@@ -1,7 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.1;net10.0</TargetFrameworks>
+    <!--
+      netstandard2.0 is required so Fallout.SourceGenerators (Roslyn) and Fallout.MSBuildTasks
+      (net472) can consume Constants. Without it those two need their own copy of the values.
+
+      Keep new Core code netstandard2.0-clean, or guard it with #if NETSTANDARD2_1_OR_GREATER.
+      Drop this target once neither of those consumers needs it.
+    -->
+    <TargetFrameworks>netstandard2.0;netstandard2.1;net10.0</TargetFrameworks>
     <Description>The reactor core of Fallout — pure domain types and graph algorithms of the build execution pipeline. No I/O, no process, no console, no logging. References nothing else in the repo.</Description>
   </PropertyGroup>
