+++ tests/Fallout.Solution.Specs/SolutionSpec.cs
@@ -10,7 +10,7 @@ namespace Fallout.Common.Specs;
 
 public class SolutionModelSpecs
 {
-    private static AbsolutePath RootDirectory => Constants.TryGetRootDirectoryFrom(EnvironmentInfo.WorkingDirectory).NotNull();
+    private static AbsolutePath RootDirectory => FalloutPaths.TryGetRootDirectoryFrom(EnvironmentInfo.WorkingDirectory).NotNull();
 
     private static AbsolutePath SolutionFile => RootDirectory / "fallout.slnx";
